Add support for vanilla.localhost/dir development
Currently, our main dev setup is for https://dev.vanilla.localhost. It works, we’ve always developed like that, and why change? The answer: URL bugs.

## The Problem

The problem with developing against the root of a domain is that we are always introducing URL bugs by not calling the correct wrapper functions on URLs and static assets respectively. These bugs happen all the time no matter how senior the developer. They also waste a lot of time. The fact of the matter is we just won’t catch them ourselves so we need to improve our tools to catch them.

With this PR I am introducing the concept of always developing against a subdirectory. With this method, instead of developing like this:

https://dev.vanilla.localhost

You would develop like this:

https://vanilla.localhost/dev

So the idea is your site always looks like it is in a subdirectory. This will catch all bugs where URLs are not wrapped in the appropriate `url()` function. Client-side URLs will also fail in development if they haven’t been properly formatted.

## Other Benefits

This PR also allows you to specify *any* directory and not just `/dev`. When specifying another directory, you will in essence be accessing a different site. This allows developers to have as many local sites as they wish.

## Bugs Not Stopped

This config will not catch bugs of static assets that are wrapped in `url()` instead of `asset()`. Hopefully, some clever other dev comes up with a solution for that.

## TODO

- [x] nginx config file
- [x] bootstrap.before.php
- [ ] SSH key
- [ ] Docs