Creator of human rights preserved
So if the governments and leaders can allow someone to issue a certain amount of pollution in the environment through a 
Environmental services ticket and their are many other  examples of dumb it down political arguments why should our youth want to be like us. I know this is silly 
But maybe we shouldn't talk about making changes and see the bigger picture in motion. There nothing a political candidate can promise that they will truly uphold because 
Debates cover such micro portion of our problems. If the devil is alive he still be tell lies just like our leaders . We can vote but don't make or break the electoral colleges vote. So I promote a boycott of the presidential election so those who get to have the vote count can be responsible for the outcomes of the election and have to answer to the public's uproar. For once we'll
See how "do as I say not as I do" is the only for virtual apes and that a parent, leader, or board member. Should and will always protect themselves and their property first...etc
 the public is not even top ten. Resources are not slim or limited but our need for money grows with are ego. God said ask and he shall receive. Money doesn't solve problems it only creates the needy & hungry  poor souls or life styles of the rich and famous. No one should have to pay to eat or have some where to sleep. We should elect a random hobo or bozo that way the can tell the youth how their future is unjustified and forsaken but don't argue it is what it is and we're all dealt the same hand you just didn't use a poker face, so we have to cattle brand you with micro chips to keep track of you and all the worthless ideas that get thrown around by technology media & social platforming talking about certain topics never killed
anyone neither did an imaginary throat punch shut me up. This might be off topic but I  speak my mind not to be heard but because it's truthfully my & our greatest gift from God. Our  voice and words came from the snake in the grass so cover up your naked... shame  and eat the fruit your father said not to for surely you will not die. Truthfully we do die but our soul wasn't ours and never up for grabs but our loyalty is confused for pride and too much to ourself only decides the future of Nothing in my hands can pull a rabbit from a hat. Magic isn't real unless your just a bitcon gullable. Replacing one lie with another is just gonna leave the world blind. Crime shouldn't never pay but payback and leaders that talk more than solving our problems shouldn't be paid anything because of net worth our belief in God and his trust the American dollar was broke. So now they scramble to find a solution when the solution has been there for us since the we first time humanity's disobedience was forgiven but at a price. ask and ye shall receive. I'm sure their is someone who disagrees. But they probably have money and everything they need or say I pray and it goes unanswered... Etc. But God didn't say ask me for it and I'll give it to ya. Funny how you always make it about you and not we.