'Readme.md'
Markdown update upgrade attach-a-pic send it to someone who cares like me. This is not a secret if it is oops! Jokes on your local Dentist or Doctor and so is the bill  for a apple a day didn't keep them away neither did the Big Mac but not everyone cant afford organically grown produce, or a fancy car, house, and clothes. Class is being turned to a level horizontal zone knowledge isn't just for kids hopefully you'll understand selfishness and greed could have ended a generation ago so it will happen. What can be done apple can lead the way to end the game of thrones with this grand gesture of dropping new updates and upgrades to all who simply ask for it or request subscription to news feeds online... Basically everyone. All the proceeds go to free health care and so forth giving. I say  advertisement can take the losses and  turn a profit for as little or next to no effort. Along with car companies as well as big tobacco, energy drinks, and news television stations (not locally).One way paying it forward, those companies will think twice before releasing marketing schemes and having terms of agreement, privacy,etc. Taking responsibility for Big Table Identity Swapping 'blacksites' 'loopholes'wormholes' and Security and government law enforcement can finally focus on protecting consumers for we always prevail if everyone pitched in there will be greater things to do than fight or no argument or pointing fingers but those baby boomers are to blame that's for sure